requirements.txt
torch>=2.5.1
torchvision>=0.20.1
diffusers>=0.32.1
transformers>=4.47.1
einops>=0.7.0
accelerate>=1.2.1
pytorch-lightning>=2.2.5
wandb>=0.16.6
PyYAML>=6.0
